disable event manager
